refact: changes PriorityQueueItem<T>::_priority data type

+++ includes/PriorityQueueItem.hpp
@@ -4,12 +4,12 @@
 template<typename T>
 class PriorityQueueItem {
     public:
-        int priority;
+        float priority;
         int index;
         T item;
 
         PriorityQueueItem(void);
-        PriorityQueueItem(int, const T&);
+        PriorityQueueItem(float, const T&);
         bool operator<(const PriorityQueueItem<T>) const;
         bool operator<=(const PriorityQueueItem<T>) const;
         bool operator>(const PriorityQueueItem<T>) const;
@@ -20,7 +20,7 @@ template<typename T>
 PriorityQueueItem<T>::PriorityQueueItem(void) {}
 
 template<typename T>
-PriorityQueueItem<T>::PriorityQueueItem(int priority, const T& item) : priority(priority), item(item), index(-1) {}
+PriorityQueueItem<T>::PriorityQueueItem(float priority, const T& item) : priority(priority), item(item), index(-1) {}
 
 template<typename T>
 bool PriorityQueueItem<T>::operator<(const PriorityQueueItem<T> rhs) const {
